requirements.txt
# Django Framework
Django==6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Web Server
gunicorn==21.2.0
whitenoise==6.6.0

# Image Processing
Pillow==10.2.0

# Environment Variables
python-decouple==3.8

# Internationalization
django-modeltranslation==0.18.12

# Admin Interface Enhancement
django-admin-interface==0.26.0

# Form Handling
django-crispy-forms==2.1
crispy-tailwind==1.0.3

# Security
django-cors-headers==4.3.1

# Timezone Support
pytz==2024.1

# Production dependencies
django-redis==5.4.0
celery==5.3.6
redis==5.0.1

# Development dependencies (optional)
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Admin Interface Enhancement
django-jazzmin

# Environment Variables
python-dotenv
